+++ protocolgame.h
@@ -163,7 +163,7 @@ class ProtocolGame : public Protocol
 		void parseViolationReport(NetworkMessage& msg);
 
 		//Send functions
-		void sendChannelMessage(std::string author, std::string text, MessageClasses type, uint8_t channel);
+		void sendChannelMessage(std::string author, std::string text, MessageClasses type, uint16_t channel);
 		void sendChannelEvent(uint16_t channelId, const std::string& playerName, ChannelEvent_t channelEvent);
 		void sendClosePrivate(uint16_t channelId);
 		void sendCreatePrivateChannel(uint16_t channelId, const std::string& channelName);
